1. Added react native elements. 2. Started wrapping up the design of the main screen.
